[Refactor] NameSpace 오류로 인한 스크립트 수정

+++ Assets/01.Scripts/Player/Player_Battle/PlayerHUD.cs
@@ -83,7 +83,7 @@ public class PlayerHUD : MonoBehaviour
         }
     }
 
-    private void UpdateHPHUD(int previous, int current)
+    private void UpdateHPHUD(float previous, float current)
     {
         if (previous - current > 0)
         {
